requirements.txt
flask
flask-cors
tensorflow==2.17.0
pillow
numpy
werkzeug
gunicorn
requests
